# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_monitoring-system_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@company.com
    name: Internal Dev Team
  description: API untuk sistem monitoring audit internal perusahaan
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Monitoring Audit API
  version: "1.0"
paths:
...
  /api/v1/uploads/{id}/status:
    get:
      consumes:
      - application/json
      description: Get the status of a specific upload
      parameters:
      - description: Upload ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get upload status
      tags:
      - uploads
  /api/v1/uploads/file:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file associated with an inspection
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: File Type
        in: formData
        name: file_type
        required: true
        type: string
      - description: Inspection ID
        in: formData
        name: inspection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - uploads
  /api/v1/uploads/inspection/{inspectionId}:
    get:
      consumes:
      - application/json
      description: Get all uploads associated with a specific inspection
      parameters:
      - description: Inspection ID
        in: path
        name: inspectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get uploads by inspection
      tags:
      - uploads
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"